Permessions by roles added

+++ app/Http/Controllers/CategoryController.php
@@ -4,17 +4,20 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Http\Requests\StoreCategoryRequest;
-use App\Http\Requests\UpdateCategoryRequest;
+
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
+    public function __construct()
+    {
+        $this->middleware('auth:api', ['except'=>['index','show']]);
+    }
+
+
     public function index()
     {
+        
         $categories = Category::orderBy('id')->get();
 
         return response()->json([
@@ -23,38 +26,24 @@ class CategoryController extends Controller
         ]); 
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \App\Http\Requests\StoreCategoryRequest  $request
-     * @return \Illuminate\Http\Response
-     */
+ 
     public function store(StoreCategoryRequest $request)
     {
-        $category = Category::create($request->all());
+        if(auth()->user()->role_id!=1){
+            return response()->json(["message"=>"This method not allowed !"]); 
+        }else{
+            $category = Category::create($request->all());
 
-        return response()->json([
-            'status' => true,
-            'message' => "Category Created successfully!",
-            'category' => $category
-        ], 201);    }
+            return response()->json([
+                'status' => true,
+                'message' => "Category Created successfully!",
+                'category' => $category
+            ], 201);   
+        } 
+    }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
+
     public function show(Category $category)
     {
         $category->find($category->id);
@@ -63,38 +52,26 @@ class CategoryController extends Controller
         }
         return response()->json($category, 200);    }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Category $category)
-    {
-        //
-    }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \App\Http\Requests\UpdateCategoryRequest  $request
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
+
+
     public function update(StoreCategoryRequest $request, Category $category)
     {
-        $category->update($request->all());
+        if(auth()->user()->role_id!=1){
+              return response()->json(["message"=>"This method not allowed !"]); 
+        }else{
+            $category->update($request->all());
 
-        if (!$category) {
-            return response()->json(['message' => 'Category not found'], 404);
+            if (!$category) {
+                return response()->json(['message' => 'Category not found'], 404);
+            }
+
+            return response()->json([
+                'status' => true,
+                'message' => "Category Updated successfully!",
+                'category' => $category,
+            ], 200);  
         }
-
-        return response()->json([
-            'status' => true,
-            'message' => "Category Updated successfully!",
-            'category' => $category
-        ], 200);   
-        
      }
 
     /**
@@ -105,16 +82,21 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
+        if(auth()->user()->role_id!=1){
+            return response()->json(["message"=>"This method not allowed !"]); 
+        }else{
+            $category->delete();
 
-        if (!$category) {
+            if (!$category) {
+                return response()->json([
+                    'message' => 'Category not found'
+                ], 404);
+            }
+
             return response()->json([
-                'message' => 'Category not found'
-            ], 404);
-        }
-
-        return response()->json([
-            'status' => true,
-            'message' => 'Category deleted successfully'
-        ], 200);    }
+                'status' => true,
+                'message' => 'Category deleted successfully'
+            ], 200);  
+        }  
+    }
 }
